# backend/api/openapi.yaml
openapi: "3.1.0"
info:
  title: VIB DevOps Ecosystem Platform — Combined API
  version: "2.0.0"
  description: |
    API contract for 4 modules added in BRD-001 v1.1 (FR-019 – FR-032).
    Auth: Bearer JWT (ADR-003). All write operations produce audit log (NFR-007).
    Banking-grade: all endpoints except /health require Authorization header.
  contact:
    name: Backend Dev (Agent 06)
servers:
  - url: http://localhost:8001
    description: PPG System — Annual Plans, Project Objects, Reports
  - url: http://localhost:8002
    description: BA Workflow — BA Documents v2
  - url: http://localhost:8003
    description: Test Platform — Test Documents
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Meta:
      type: object
      properties:
        total:
          type: integer
        page:
          type: integer
        size:
          type: integer
    AnnualPlan:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        year:
          type: integer
          minimum: 2020
          maximum: 2050
        description:
          type: string
          nullable: true
        status:
          type: string
          enum: [draft, active, closed]
        objectives_count:
          type: integer
        dod_completion_pct:
          type: number
          format: float
        projects_count:
          type: integer
        created_at:
          type: string
          format: date-time
        created_by:
          type: string
    AnnualPlanCreate:
      type: object
      required: [name, year, objectives]
      properties:
        name:
          type: string
          maxLength: 300
        year:
          type: integer
          minimum: 2020
          maximum: 2050
        description:
          type: string
          nullable: true
        objectives:
          type: array
          minItems: 1
          items:
            type: object
            required: [title]
            properties:
              title:
                type: string
                maxLength: 500
              description:
                type: string
                nullable: true
              sort_order:
                type: integer
                default: 0
        dod_items:
          type: array
          items:
            type: object
            required: [criterion]
            properties:
              criterion:
                type: string
                maxLength: 500
              weight:
                type: number
                default: 1.0
                exclusiveMinimum: 0
                maximum: 100
...
security:
  - BearerAuth: []
paths:
...
  /api/v1/annual-plans:
    get:
      summary: List annual plans (FR-019)
      operationId: listAnnualPlans
      tags: [annual-plans]
      parameters:
        - name: year
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [draft, active, closed]
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: size
          in: query
          schema:
            type: integer
            default: 20
            maximum: 100
      responses:
        "200":
          description: List of annual plans
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: "#/components/schemas/AnnualPlan"
                  meta:
                    $ref: "#/components/schemas/Meta"
        "401":
          description: Unauthorized

    post:
      summary: Create annual plan (FR-019)
      operationId: createAnnualPlan
      tags: [annual-plans]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AnnualPlanCreate"
      responses:
        "201":
          description: Created annual plan
        "400":
          description: Missing objectives
        "422":
          description: Validation error
  /api/v1/annual-plans/{plan_id}:
    get:
      summary: Get annual plan details (FR-019)
      operationId: getAnnualPlan
      tags: [annual-plans]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: Annual plan with objectives, dod_items, projects
        "404":
          description: Not found

    put:
      summary: Update plan name/description
      operationId: updateAnnualPlan
      tags: [annual-plans]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
                description:
                  type: string
      responses:
        "200":
          description: Updated plan

    delete:
      summary: Soft delete plan (draft only)
      operationId: deleteAnnualPlan
      tags: [annual-plans]
      parameters:
        - name: plan_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "204":
          description: Deleted
        "409":
          description: Plan is active or closed
...
  /api/v1/documents:
    get:
      summary: List BA documents (FR-027, FR-028, FR-029)
      operationId: listBADocuments
      tags: [ba-documents]
      parameters:
        - name: project_id
          in: query
          schema:
            type: string
            format: uuid
        - name: doc_type
          in: query
          schema:
            type: string
            enum: [BRD, BRS, FSD, API_SPEC, ERD, DATA_DICT, WIREFRAME, PROCESS_FLOW]
        - name: status
          in: query
          schema:
            type: string
        - name: object_id
          in: query
          schema:
            type: string
            format: uuid
          description: Filter by linked object (FR-027)
        - name: milestone_id
          in: query
          schema:
            type: string
            format: uuid
      responses:
        "200":
          description: List of BA documents

    post:
      summary: Create BA document (FR-027, FR-029)
      operationId: createBADocument
      tags: [ba-documents]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [project_id, doc_type, title]
              properties:
                project_id:
                  type: string
                  format: uuid
                doc_type:
                  type: string
                  enum: [BRD, BRS, FSD, API_SPEC, ERD, DATA_DICT, WIREFRAME, PROCESS_FLOW]
                title:
                  type: string
                  maxLength: 300
                content:
                  type: string
                  default: ""
                milestone_id:
                  type: string
                  format: uuid
                  nullable: true
                object_ids:
                  type: array
                  items:
                    type: string
                    format: uuid
                metadata:
                  type: object
      responses:
        "201":
          description: Created document
        "422":
          description: OBJECT_INCOMPLETE or validation error (BR-011)
...